wiki-ai/pyproject.toml
[project]
name = "wiki-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub (>=2.6.1,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "plotly (>=6.1.1,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "hiredis (>=3.2.1,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "wiki_ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
